git-pr: move subcommands to separate files
